refactor: 클래스 네임 수정
