+++ src/test/java/domain/card/CardDeckTest.java
@@ -1,64 +1,54 @@
 package domain.card;
 
 import static org.assertj.core.api.Assertions.assertThat;
+import static org.junit.jupiter.api.Assertions.assertAll;
 
 import java.util.List;
+
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator.ReplaceUnderscores;
 import org.junit.jupiter.api.Test;
 
 @DisplayNameGeneration(ReplaceUnderscores.class)
-class CardDeckTest {
-
-    private final CardDeck cardDeck = new CardDeck();
-    private final Card jClover = new Card(Rank.JACK, Shape.CLOVER);
-    private final Card fiveHeart = new Card(Rank.FIVE, Shape.HEART);
-    private final Card aceHeart = new Card(Rank.ACE, Shape.HEART);
-    private final Card aceClover = new Card(Rank.ACE, Shape.CLOVER);
-    private final Card twoDiamond = new Card(Rank.TWO, Shape.DIAMOND);
+public class CardDeckTest {
 
     @Test
-    void 카드덱의_초기카드_2장을_설정할수있다() {
-        cardDeck.setUpCards(jClover, fiveHeart);
+    void 카드팩은_Rank와_Shape의_모든_조합의_카드를_가지고_있다() {
+        CardDeck pack = new CardDeck();
 
-        List<Card> cards = cardDeck.getCards();
-        assertThat(cards).contains(jClover, fiveHeart);
+        var cards = pack.getCards();
+
+        assertCardPackHasAllCombinations(cards);
     }
 
     @Test
-    void 카드의_점수의_합을_구할수있다() {
-        cardDeck.setUpCards(jClover, fiveHeart);
+    void 카드팩을_섞을때_모든_조합을_가지고있다() {
+        CardDeck pack = new CardDeck();
 
-        var score = cardDeck.calculateScore();
+        pack.shuffle();
+        var cards = pack.getCards();
 
-        assertThat(score).isEqualTo(15);
+        assertCardPackHasAllCombinations(cards);
+    }
+
+    private static void assertCardPackHasAllCombinations(List<Card> cards) {
+        for (Rank rank : Rank.values()) {
+            for (Shape shape : Shape.values()) {
+                Card card = new Card(rank, shape);
+                assertThat(cards).contains(card);
+            }
+        }
     }
 
     @Test
-    void A가있을때_나머지숫자의_합이_11이상이면_1을_선택() {
-        cardDeck.setUpCards(aceHeart, aceClover);
+    void 카드팩_상단의_한장을_꺼낸다() {
+        CardDeck pack = new CardDeck();
+        Card card = pack.poll();
 
-        var score = cardDeck.calculateScore();
-
-        assertThat(score).isEqualTo(12);
-    }
-
-    @Test
-    void A가있을때_나머지숫자의_합이_10이하이면_11을_선택() {
-        cardDeck.setUpCards(aceHeart, jClover);
-
-        var score = cardDeck.calculateScore();
-
-        assertThat(score).isEqualTo(21);
-    }
-
-    @Test
-    void 카드덱에_카드를_추가한다() {
-        cardDeck.setUpCards(jClover, fiveHeart);
-
-        cardDeck.takeMore(twoDiamond);
-
-        List<Card> cards = cardDeck.getCards();
-        assertThat(cards).contains(jClover, fiveHeart, twoDiamond);
+        assertAll(
+                () -> assertThat(card).isNotNull(),
+                () -> assertThat(pack.getCards()).doesNotContain(card),
+                () -> assertThat(card).isEqualTo(new Card(Rank.ACE, Shape.SPADE))
+        );
     }
 }
